Reordered code and worked on Operator's order of operations
Also improved formatting and removed an unneeded header file corresponding to calc.cpp (calc.h).

I created a tests.bat, but it is really just something that I quickly threw together to test a few things; its path to calc.exe doesn't match up with the build instructions in the README, so I'm excluding it from the commits.